+++ src/components/rainViewerData.js
@@ -1,20 +1,22 @@
 import axios from "axios";
-import { useEffect, useState } from "react";
 
 const baseURL = "https://api.rainviewer.com/public/weather-maps.json";
 
 let url = '';
 let path = '';
 
+let fullURL = '';
+
+
+
 export const GetRainViewerData = () => {
 
   axios.get(baseURL)
   .then(function (response) {
-    //console.log(response);
+    const lastFrame = response.data.radar.past.length-1;
     url = response.data.host;
-    path = response.data.radar.past;
-    console.log(url);
-    console.log(path);
+    path = response.data.radar.past[lastFrame].path;
+    fullURL = url + path + '/8000/2/0_1.png';
   })
   .catch(function (error) {
     console.log(error);
@@ -22,4 +24,6 @@ export const GetRainViewerData = () => {
   .then(function () {
     // always executed
   });  
+
+  return fullURL;
 };